requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
Babel==2.10.1
backcall==0.2.0
boto==2.49.0
cachetools==4.2.4
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
cloud-tpu-client==0.10
colorama==0.4.4
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
dill==0.3.5.1
distro==1.7.0
editdistance==0.6.0
entrypoints==0.4
etils==0.6.0
filelock==3.7.1
flatbuffers==1.12
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.5.0
future==0.18.2
gast==0.4.0
gcsfs==2022.5.0
gevent==21.12.0
gin-config==0.5.0
google-api-core==1.31.6
google-api-python-client==1.8.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
google-compute-engine==2.8.13
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
greenlet==1.1.2
grpcio==1.46.3
h5py==3.7.0
httplib2==0.20.4
huggingface-hub==0.7.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
ipykernel==6.13.0
ipython==7.34.0
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.3.1
jupyter-core==4.10.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
Markdown==3.3.7
matplotlib==3.5.2
matplotlib-inline==0.1.3
mesh-tensorflow==0.1.21
multidict==6.0.2
nest-asyncio==1.5.5
nltk==3.7
numpy==1.21.6
oauth2client==4.1.3
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
portalocker==2.4.0
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocoevalcap @ git+https://github.com/salaniz/pycocoevalcap@ad63453cfab57a81a02b2949b17a91fab1c3df77
pycocotools==2.0.4
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==23.0.0
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
rouge-score==0.0.4
rsa==4.8
sacrebleu==2.1.0
scikit-learn==1.0.2
scipy==1.7.3
sentencepiece==0.1.96
seqio==0.0.7
six==1.16.0
t5==0.9.3
tabulate==0.8.9
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-datasets==4.5.2
tensorflow-estimator==2.9.0
tensorflow-gcs-config==2.9.1
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-metadata==1.8.0
tensorflow-text==2.9.0
termcolor==1.1.0
tfds-nightly==4.5.2.dev202206010819
threadpoolctl==3.1.0
tokenizers==0.12.1
toml==0.10.2
torch==1.11.0
tornado==6.1
tqdm==4.64.0
traitlets==5.2.2.post1
transformers==4.19.2
typing_extensions==4.2.0
uritemplate==3.0.1
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
